athena_tool/requirements.txt
# ATHENA Python Package Dependencies
click~=7.1.2
halo~=0.0.31
tabulate~=0.8.9
pyaml~=20.4.0
config~=0.5.0
utils~=1.0.1
configparser~=5.0.2
bidict~=0.21.2
numpy~=1.19.2
matplotlib~=3.3.4
scipy~=1.6.0
pyyaml~=5.4.1
pandas~=1.2.1
natsort~=7.1.1
colorama~=0.4.4
six~=1.15.0
pytz~=2021.1
pillow~=8.1.0
tornado~=6.1
python-dateutil~=2.8.1
pyparsing~=2.4.7 
pyfiglet~=0.8.post1
yamlordereddictloader~=0.4.0
typer~=0.3.2
setuptools-scm~=6.0
cffi~=1.14.6
importlib-metadata~=4.6.3
tqdm~=4.0
keras~=2.4
pyscaffold~=4.0
pytest~=6.2
pytest-cov~=2.12
tox~=3.23
sphinx~=4.0
python-dotenv~=0.18
datadict~=0.1.0
torch==1.8.1+cpu
torchvision==0.9.1+cpu
torchaudio==0.8.1
